+++ lib/synchronization/binary_semaphore.dart
@@ -2,24 +2,43 @@ import 'dart:async';
 
 import 'package:meta/meta.dart';
 
-import 'counting_semaphore.dart';
+import '../src/synchronization/wait_queue.dart';
 import 'synchronizer.dart';
 
 /// A [BinarySemaphore] is a synchronization primitive with an integer value
 /// restricted to 0 or 1, representing locked (0) or unlocked (1) states.
 class BinarySemaphore implements Synchronizer {
-  final CountingSemaphore _lock = CountingSemaphore(0, 1);
+  static final Future<bool> _true = Future.value(true);
+
+  static final Future<void> _void = Future.value();
+
+  bool _isLocked = false;
+
+  final WaitQueue _waitQueue = WaitQueue();
 
   /// Acquires a permit from this semaphore.
   @override
   Future<void> acquire() {
-    return _lock.acquire();
+    if (!_isLocked) {
+      _isLocked = true;
+      return _void;
+    }
+
+    return _waitQueue.enqueue();
   }
 
   /// Releases a permit.
   @override
   Future<void> release() {
-    return _lock.release();
+    if (!_isLocked) {
+      throw StateError('Unmatched call of \'release()()\' method');
+    }
+
+    if (!_waitQueue.dequeue()) {
+      _isLocked = false;
+    }
+
+    return _void;
   }
 
   /// Tries to acquire a permit from this semaphore and waits until the
@@ -32,6 +51,16 @@ class BinarySemaphore implements Synchronizer {
   @override
   @useResult
   Future<bool> tryAcquire(Duration timeout) {
-    return _lock.tryAcquire(timeout);
+    if (timeout.isNegative) {
+      throw ArgumentError.value(
+          timeout, 'timeout', 'Timeout must not be negative');
+    }
+
+    if (!_isLocked) {
+      _isLocked = true;
+      return _true;
+    }
+
+    return _waitQueue.enqueue(timeout);
   }
 }