+++ lib/synchronization/reentrant_lock.dart
@@ -1,9 +1,13 @@
 import 'package:meta/meta.dart';
 
 import '../src/task/task.dart';
-import 'binary_semaphore.dart';
+import 'counting_semaphore.dart';
 import 'synchronizer.dart';
 
+/// A [ReentrantLock]  is a synchronization primitive that works like a mutex.\
+/// It blocks execution of all tasks that do not own this lock.\
+/// The task that acquired the permit becomes the owner of this lock.\
+/// The task owner can enter and exit as long as it holds this lock.
 class ReentrantLock implements Synchronizer {
   static final Future<void> _void = Future.value();
 
@@ -11,7 +15,7 @@ class ReentrantLock implements Synchronizer {
 
   AnyTask? _owner;
 
-  final BinarySemaphore _sem = BinarySemaphore();
+  final CountingSemaphore _sem = CountingSemaphore(0, 1);
 
   @override
   Future<void> acquire() async {
@@ -35,12 +39,13 @@ class ReentrantLock implements Synchronizer {
   @override
   Future<void> release() {
     if (_owner == null) {
-      throw StateError('Attempting to release a mutex that is unowned');
+      throw StateError(
+          'Attempting to release a reentrant lock that is unowned');
     }
 
     if (_owner != Task.current) {
       throw StateError(
-          'Attempting to release a mutex that is owned by a different object');
+          'Attempting to release a reentrant lock that is owned by a different task');
     }
 
     if (--_count > 0) {